requirements.txt
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain>=0.1.0
langgraph>=0.0.20
chromadb>=0.4.0
pydantic>=2.0.0
streamlit
pandas
altair
mcp>=1.0.0
authlib
requests-oauthlib
feedparser
requests
python-dotenv
tenacity>=8.0.0
rank-bm25>=0.2.2
jieba>=0.42
